Made game more modular and removed terminal input
